feat(search): Added click on category button and set query in input field
