+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Extensions/SearchResultsView.extension.swift
@@ -10,9 +10,4 @@ extension SearchResultsView: UITableViewDelegate, UITableViewDataSource {
         cell.configure(with: results[indexPath.row])
         return cell
     }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let selectedResult = results[indexPath.row]
-        delegate?.didSelectResult(selectedResult)
-    }
 }
