+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Model/SearchWidgetCategoriesButtonType.enum.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+public enum SearchWidgetCategoriesButtonType {
+    case bordered
+    case blacked
+    case colored
+}
