+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Model/SearchWidgetObjectificationType.enum.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum SearchWidgetObjectificationType {
+    case properties
+    case values
+    case all
+}
