--- REES46/Classes/Stories/StoryViewController/SearchWidget/Protocol/SearchResultsViewDelegate.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-public protocol SearchResultsViewDelegate: AnyObject {
-     func didSelectResult(_ result: SearchResult)
-}
